requirements.txt
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mysqlclient==2.0.1
Flask-SQLAlchemy==2.4.2
Flask-Cors==3.0.8
SQLAlchemy==1.3.19
python-dateutil==2.8.1
requests
